requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
openai>=1.0.0
asyncio-timeout>=4.0.0
argparse 